model_manager/requirements.txt
fastapi==0.108.0
uvicorn==0.20.0
sqlalchemy
pydantic==1.10.4
python-multipart==0.0.5
aiofiles==22.1.0
python-dotenv==0.21.1
psycopg2-binary==2.9.6
python-jose==3.3.0
bcrypt==4.0.1
passlib==1.7.4
alembic
minio==7.1.13
packaging==23.2
pamqp==3.2.1
loguru
asyncpg
websockets
requests
celery
redis
fastapi-utils

sentry-sdk[fastapi]
Pillow

gunicorn
pymongo
motor
httpx

pandas
scikit-learn
